refactor(ldes): put configuration in more explicitly named file
The contents of, the new, `config.ts` combines configuration used for
initialisation, dispatch, and healing.  Putting this in a designated config file
that is read by these different flows makes this more explicit.

The previous situation with this config in `initialization.ts` can cause
confusion. For example, why and how the `healingPredicates` would be used during initialisation.
